changed all objects to classes (better looking objects). Still need to figure out module exports though.
