+++ app/index.html
@@ -35,10 +35,7 @@
     </div>
 
     <!-- Javascript -->
-    <script src='./js/init.js'></script>
-    <script src='./js/buttons.js'></script>
-    <script src='./js/directory.js'></script>
-    <script src='./js/melodii.js'></script>
+    <script src='./js/main.js'></script>
 </body>
 
 </html>
